20241214
manager&report

+++ src/main/java/com/example/demo/service/Impl/ReportServiceImpl.java
@@ -1,16 +1,19 @@
 package com.example.demo.service.Impl;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.example.demo.exception.UserNotFoundException;
+import com.example.demo.exception.ReportNotFoundException;
 import com.example.demo.mapper.Mapper;
 import com.example.demo.model.dto.ReportListDto;
 import com.example.demo.model.entity.ReportList;
+import com.example.demo.repository.LovehomeRepository;
 import com.example.demo.repository.ReportRepository;
+import com.example.demo.repository.UserRepository;
 import com.example.demo.service.ReportService;
 
 @Service
@@ -20,12 +23,19 @@ public class ReportServiceImpl implements ReportService {
 	private ReportRepository reportRepository;
 	
 	@Autowired
-	private Mapper mapper;
+	private UserRepository userRepository;
 	
+	@Autowired
+	private LovehomeRepository lovehomeRepository;
+	
+	@Autowired
+	private Mapper mapper;
 	
 	@Override
 	public ReportListDto addReport(ReportListDto reportDto) {
 		ReportList reportList = mapper.toReportListEntity(reportDto);
+		reportList.setUser(userRepository.findById(reportDto.getUserDto().getUserId()).get());
+		reportList.setLovehome(lovehomeRepository.findById(reportDto.getLovehomeId()).get());
 		reportRepository.save(reportList);
 		return mapper.toReportListDto(reportList);
 	}
@@ -38,25 +48,37 @@ public class ReportServiceImpl implements ReportService {
 	@Override
 	public ReportListDto getReportByNumber(Integer reportNumber) {
 		ReportList report = reportRepository.findById(reportNumber)
-				.orElseThrow(() -> new UserNotFoundException("無此通報單：" + reportNumber));
+				.orElseThrow(() -> new ReportNotFoundException("無此通報表單：" + reportNumber));
 		return mapper.toReportListDto(report);
 	}
 
 	@Override
 	public ReportListDto updateReport(ReportListDto reportDto) {
-		return reportRepository.findById(reportDto.getReportNumber())
-				.map(report ->{
-					mapper.toReportListEntity(reportDto);
-					ReportList upReportList = reportRepository.save(report);
-					return mapper.toReportListDto(upReportList);
-				})
-				.orElseThrow(() -> new RuntimeException("查無資料"));
+		ReportList updateReportList = reportRepository.findById(reportDto.getReportNumber()).get();
+		updateReportList.setReportStatus(reportDto.getReportStatus());
+		reportRepository.save(updateReportList);
+		return mapper.toReportListDto(updateReportList);
 	}
 
 	@Override
 	public void deleteReport(Integer reportNumber) {
+		Optional<ReportList> optReportList = reportRepository.findById(reportNumber);
+		if (optReportList.isEmpty()) {
+			throw new ReportNotFoundException("無此通報清單:" + reportNumber);
+		}
 		reportRepository.deleteById(reportNumber);
-		
 	}
 
+	@Override
+	public List<ReportListDto> getReportByUserId(Integer userId) {
+		return reportRepository.findByUserUserId(userId).stream()
+				.map(mapper::toReportListDto).collect(Collectors.toList());
+	}
+
+	@Override
+	public List<ReportListDto> getReportByLovehomeId(Integer lovehomeId) {
+		List<ReportListDto> reportListDtos = reportRepository.findByLovehomeLovehomeId(lovehomeId).stream()
+				.map(mapper::toReportListDto).collect(Collectors.toList());
+		return reportListDtos;
+	}
 }
